New draft of chat sidebar
